Prepare dlls/{comctl32,gdi,msvideo,setupapi,shell32,twain,winmm} for the conversion of HWND to a void*.

+++ dlls/setupapi/queue.c
@@ -955,7 +955,7 @@ BOOL WINAPI SetupCommitFileQueueW( HWND owner, HSPFILEQ handle, PSP_FILE_CALLBAC
     if (!queue->copy_queue.count && !queue->delete_queue.count && !queue->rename_queue.count)
         return TRUE;  /* nothing to do */
 
-    if (!handler( context, SPFILENOTIFY_STARTQUEUE, owner, 0 )) return FALSE;
+    if (!handler( context, SPFILENOTIFY_STARTQUEUE, (UINT)owner, 0 )) return FALSE;
 
     /* perform deletes */
